Handle missing series + topics when rendering shortcode
Depending on whether a "series" or "video" attribute are passed to the [wp101] shortcode (with precedence being given to the first), the shortcode will render either a playlist or a single video.

Unfortunately, if the given series and/or video doesn't exist, the shortcode was previously generating an error. This commit ensures that we're explicitly checking for values before attempting to display them, and showing a debug message to logged-in users who might need to troubleshoot.