Re-organize the WP101 settings page
Instead of always displaying the form, show the stored API key (which will soon be masked) if one exists _unless_ the user asks for the form.

This gets around the issue of the WordPress Settings API erasing the stored API key if the value submitted is empty *and* let us avoid toggling between text and password inputs, as the latter would likely cause bug reports due to Chrome (and others) warning that pages with <input type="password" /> served over HTTP are insecure.