Don't tell someone about something they've already subscribed to
The wp101-available-series option is only updated when a plugin is activated/deactivated, but if a user has subscribed to an add-on *without* changing their active_plugins list, they should still be relieved of seeing the "Would you like to add the tutorial videos for {plugin} from WP101?" notification.

Fixes #70.