Add a new wp101_is_relevant_series filter
The WP101 plugin automatically attempts to filter out series that are not relevant for a given site (e.g. if a site isn't running Jetpack, there's no reason to show the Jetpack videos). However, sometimes it's desired to override this behavior, so the new `wp101_is_relevant_series` filter is here to accomplish that.

Usage example:

    add_filter( 'wp101_is_relevant_series', function ( $is_relevant, $series ) {
        return 'some-series' === $series['slug'] ? true : $is_relevant;
    } );