Send the Origin, not X-Forwarded-Host, HTTP header with API requests
Within the WP101 Plugin app, we need to know what domain the plugin is running on so that we may ensure the correct videos are getting served.

For the longest time, we've been using `X-Forwarded-Host`, but this header will get overwritten on systems leveraging any sort of load balancer or other proxy. The **proper** header in this case is `Origin`. [From MDN](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Origin):

> The `Origin` request header indicates the origin (scheme, hostname, and port) that caused the request. For example, if a user agent needs to request resources included in a page, or fetched by scripts that it executes, then the origin of the page may be included in the request.